backend/requirements.txt
asgiref==3.6.0
bcrypt==4.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-cors-headers==4.0.0
django-environ==0.10.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
numpy==1.24.3
pandas==2.0.1
phonenumbers==8.13.11
psycopg2==2.9.6
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
